feat: enhance community dashboard with new components and data fetching
* Refactor CommunityPage to use async/await for data fetching and improve loading state management.
* Introduce CommunityFeed, CommunityMembers, and CommunityInfo components for better organization and modularity.
* Update dashboard layout to include member count and community information.
* Remove unused code and optimize data fetching logic.
