+++ app/dashboard/page.tsx
@@ -4,13 +4,14 @@ import Link from "next/link";
 export default function DashboardPage() {
   return (
     <div>
-      <h1>Dashboard</h1>
-      <Button>
-        <Link href="/dashboard/settings">Settings</Link>
-      </Button>
-      <Button>
-        <Link href="/dashboard/communities">Communities</Link>
-      </Button>
+      <div className="flex gap-2">
+        <Button>
+          <Link href="/dashboard/settings">Settings</Link>
+        </Button>
+        <Button>
+          <Link href="/dashboard/communities">Communities</Link>
+        </Button>
+      </div>
     </div>
   );
 }
